Search LILITH_PATH env var when loading files
LILITH_PATH is a colon separated list of paths to search through when
loading lilith code files. The local path is searched first, if not
found each entry in LILITH_PATH is searched through in turn.

The top-level Makefile uses this mechanism so that the unit test file
doesn't need to know where stdlib.llth lives.