Implement reader to build lval from tokens
The tokeniser produces a stream of tokens from an input string. The
reader builds up a hierarchy of lvals from the token stream and returns
the root element.

The tokeniser is also responsible for identifying the type of the token.
It is much easier to do this here rather than in the reader. One pass
over the input string should be sufficient.

The tokeniser keeps track of the line number and character position as
it copies characters from the input string to the next buffer.

Remove MPC parser from the REPL, load and read functions