Add 'try' function for basic error handling
Remove error checking code from lval_eval_sexpr() and implement a macro
to check for errors in argument lists. The macro can be called by
functions which do not expect errors and the error results will float up
the expression tree as before. The builtin_try function can now be
written to evaluate a q-expression if its first param is an error.