Replace lenv arrays with hash table
Built-in functions should not be re-bound. Previously this was managed
with a flag on the value within the lenv but this commit changes it to a
flag on the lenv itself.

lenv_copy() cannot use the clxns_copy() function because it needs to
perform a deep copy of the contents of the lenv.