Improve local bindings
Replace Lilith defined let/do/put combo to define a local binding. Let
is now defined in C and works similarly to def, but an additional
q-expression is passed in at the end of the arg list. This is the
expression to be evaluated with the local bindings passed in to a new
environment.